Update function to account for rotated grids
